refactor conditionals into more files, add one test as a goal
