+++ lib/src/modules/conditionals/ssa.dart
@@ -1,0 +1,16 @@
+import 'package:rohd/rohd.dart';
+
+/// A signal that represents an SSA node in [Combinational.ssa] which is
+/// associated with one specific [Combinational].
+class SsaLogic extends Logic {
+  /// The signal that this represents.
+  final Logic ref;
+
+  /// A unique identifier for the context of which [Combinational.ssa] it is
+  /// associated with.
+  final int context;
+
+  /// Constructs a new SSA node referring to a signal in a specific context.
+  SsaLogic(this.ref, this.context)
+      : super(width: ref.width, name: ref.name, naming: Naming.mergeable);
+}
